+++ ShoppingCart.Domain/Entities/User.cs
@@ -3,7 +3,7 @@
 public class User
 {
     public int Id { get; set; }
-    public required double Dni { get; set; }
+    public required long Dni { get; set; }
     public required string Name { get; set; }
     public required bool IsVip { get; set; }
 }
